+++ src/num/mod.rs
@@ -1,7 +1,7 @@
 // devela::num
 //
 //! Numerical types and math operations.
-#![doc = crate::doc_!(modules: crate; num: alg, geom, logic, measure, niche, rand, wave)]
+#![doc = crate::doc_!(modules: crate; num: alg, geom, logic, measure, niche, ord, rand, wave)]
 #![doc = crate::doc_!(newline)]
 //!
 #![doc = crate::doc_!(extends: cmp, num)]
@@ -16,7 +16,6 @@ mod float; // fsize, ExtFloat, ExtFloatConst, Float
 mod frac; // Frac
 mod int; // [i|u]size_[down|up], Int
 mod no; // NoNum
-mod ord; // Compare
 mod primitive; // Cast, Primitive[Cast|Join|Split]
 mod sign; // Sign
 mod traits; // Num, NumRef
@@ -27,8 +26,9 @@ mod unit; // Unit, Unit[Bi|Si]
 
 pub mod geom;
 pub mod logic;
-pub mod measure; // Interval, Cycle, Spacing…
-pub mod niche;
+pub mod measure; // Cycle*, Interval
+pub mod niche; // NonZero*, NonValue*, NonRange*
+pub mod ord; // Compare
 pub mod rand;
 
 #[cfg(feature = "alg")]
@@ -42,19 +42,20 @@ crate::items! { // structural access: _mods, _pub_mods, _internals, _all, _alway
     #[allow(unused)]
     pub use {_internals::*, _mods::*};
     #[allow(unused)] #[doc(hidden, no_inline)]
-    pub use {_always::*, _pub_mods::*,};
+    pub use {_always::*, _pub_mods::*};
 
     mod _mods { #![allow(unused)]
         pub use super::{
             error::*, float::_all::*, frac::_all::*, int::_all::*,
-            no::*, ord::_all::*, primitive::_all::*, sign::*, traits::*,
+            no::*, primitive::_all::*, sign::*, traits::*,
         };
         #[cfg(feature = "unit")]
         pub use super::unit::_all::*;
     }
     mod _pub_mods {
         pub use super::{
-            geom::_all::*, logic::_all::*, measure::_all::*, niche::_all::*, rand::_all::*,
+            geom::_all::*, logic::_all::*, measure::_all::*,
+            niche::_all::*, ord::_all::*, rand::_all::*,
         };
 
         #[cfg(feature = "alg")]